==> 02_create_data(texts_similarity).ipynb <==
import json
with open("stair_captions/stair_captions_v1.2_val.json", "r",encoding="utf-8") as f:
  json_data_val = json.load(f)
with open("stair_captions/stair_captions_v1.2_train.json", "r",encoding="utf-8") as f:
  json_data_train = json.load(f)

dataset = {}
ids = []
captions = []

def build_dataset(dataset, json_data):
  num_samples = len(json_data['annotations'])
  for i in range(num_samples):
    anno = json_data['annotations'][i]
    image_id = anno["image_id"]
    image_captions = dataset.get(image_id, [])
    image_captions.append((anno["id"], anno["caption"]))
    ids.append(anno["id"])
    captions.append(anno["caption"])
    dataset[image_id] = image_captions

build_dataset(dataset, json_data_val)
build_dataset(dataset, json_data_train)

#dataset

id2idx = {id:idx for idx, id in enumerate(ids)}
#id2idx

len(captions)

import numpy as np
import spacy
import pkg_resources, imp
imp.reload(pkg_resources)

#nlp = spacy.load('ja_ginza_nopn')
nlp = spacy.load("ja_ginza")

vectors = []
i = 0
for caption in captions:
    if i %1000 ==0:
        print(i)
    doc = nlp(caption, disable=['ner'])
    vectors.append(doc.vector)
    i = i +1
'''
'''
'''
import multiprocessing as mp
pool = mp.Pool(mp.cpu_count())
def count_sim(caption):
    #if i %1000 ==0:
    print(i)
    doc = nlp(caption, disable=['ner'])
    i = i +1
    return doc.vector
results = [pool.apply(count_sim, args=(caption)) for caption in captions]
#results = [pool.map(count_sim,captions)]  
'''

del nlp

from sklearn.metrics.pairwise import cosine_similarity
def similarity(id1, id2):
  return cosine_similarity([vectors[id2idx[id1]]], [vectors[id2idx[id2]]])[0][0]

import random

def make_triplets(dataset, threshold=0.85, seed=7, max_tries=25):
  triplets = []
  random.seed(seed)
  neg_candidate_indices = list(range(len(ids)))
  random.shuffle(neg_candidate_indices)
  def log(i, str):
    if i % 5000 == 0:
      print(str)

  for i, image_id in enumerate(list(dataset.keys())): 
    log(i, "### %d ###" % (image_id))

    # pickup positive pair.
    score = 0.0
    tries = 0
    while score < threshold : 
      [(id, caption), (id_pos, caption_pos)] = random.sample(dataset[image_id],2)
      score = similarity(id, id_pos)
      tries+=1
      if tries > max_tries:
        break
    if score < threshold:
      continue

    # pickup negative one.
    id_neg = id
    current_caption_ids = [id_cap[0] for id_cap in dataset[image_id]]
    while id_neg in current_caption_ids:
      idx_neg = neg_candidate_indices.pop()
      id_neg = ids[idx_neg]
    caption_neg = captions[id2idx[id_neg]]

    log(i, "  pos:  score: %4.2f [%s]:[%s]" % (score, caption, caption_pos))
    log(i, "  neg:  score: %4.2f [%s]:[%s]" % (similarity(id, id_neg), caption, caption_neg))
    triplets.append({
      "image_id": image_id,
      "id": id,
      "id_pos": id_pos,
      "id_neg": id_neg,
      "caption": caption,
      "caption_pos": caption_pos,
      "caption_neg": caption_neg  
    })
  return triplets

triplets = make_triplets(dataset)

len(triplets)

import pandas as pd
dt = pd.DataFrame(triplets)
dt

import joblib
joblib.dump(dt,"texts_orinal_data.pkl")

joblib.load("texts_orinal_data.pkl")
