==> 03_analysis_data(texts_similarity).ipynb <==
import pickle
from transformers import BertJapaneseTokenizer, BertForMaskedLM
import pandas as pd
import json
import matplotlib.pyplot as plt
import tensorflow as tf
import joblib

import seaborn as sns
import collections

df = joblib.load("texts_orinal_data.pkl")
df = df[["caption","caption_pos","caption_neg"]]
def make_l(v):
    return len(v)
for col in ["caption","caption_pos","caption_neg"]:
    df[col+"_L"] = df[col].apply(make_l)

df["caption_L"].describe(percentiles=[0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1.0])

df["caption_pos_L"].describe(percentiles=[0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1.0])

df["caption_neg_L"].describe(percentiles=[0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1.0])

from transformers import DistilBertTokenizer
max_seq_length = 28
train_output_file = "texts_train_data.tfrecord"
test_output_file = "texts_test_data.tfrecord"
tokenizer = DistilBertTokenizer.from_pretrained("cl-tohoku/bert-base-japanese-whole-word-masking")

def create_code(text1,text2,label_id,tf_record_writer):
    all_features = collections.OrderedDict()
    #encode_plus_tokens = tokenizer.encode_plus(f"{text1}[SEP]{text2}",padding='max_length',max_length=max_seq_length*2,truncation=True)
    encode_plus_tokens = tokenizer.encode_plus(text1,padding='max_length',max_length=max_seq_length,truncation=True)
    input_ids1 = encode_plus_tokens["input_ids"]
    input_mask1 = encode_plus_tokens["attention_mask"]
    segment_ids1 = [0] *max_seq_length
    
    encode_plus_tokens = tokenizer.encode_plus(text2,padding='max_length',max_length=max_seq_length,truncation=True)
    input_ids2 = encode_plus_tokens["input_ids"]
    input_mask2 = encode_plus_tokens["attention_mask"]
    segment_ids2 = [0] *max_seq_length

    #label_id = text = dt["Y"][i]
    all_features["input_ids1"] = tf.train.Feature(int64_list=tf.train.Int64List(value=input_ids1))
    all_features["input_mask1"] = tf.train.Feature(int64_list=tf.train.Int64List(value=input_mask1))
    all_features["segment_ids1"] = tf.train.Feature(int64_list=tf.train.Int64List(value=segment_ids1))
    all_features["input_ids2"] = tf.train.Feature(int64_list=tf.train.Int64List(value=input_ids2))
    all_features["input_mask2"] = tf.train.Feature(int64_list=tf.train.Int64List(value=input_mask2))
    all_features["segment_ids2"] = tf.train.Feature(int64_list=tf.train.Int64List(value=segment_ids2))
    all_features["label_ids"] = tf.train.Feature(int64_list=tf.train.Int64List(value=[label_id]))
    tf_record = tf.train.Example(features=tf.train.Features(feature=all_features))
    tf_record_writer.write(tf_record.SerializeToString())
def create_tf_data_file(dt,output_file):
    tf_record_writer = tf.io.TFRecordWriter(output_file)
    for i in range(dt.shape[0]):
        #label_id = i%2
        text1 = dt["caption"][i][0:max_seq_length]
        text2 = dt["caption_pos"][i][0:max_seq_length]
        text3 = dt["caption_neg"][i][0:max_seq_length]
        create_code(text1,text2,1,tf_record_writer)
        create_code(text1,text3,0,tf_record_writer)
    tf_record_writer.close()

from sklearn.model_selection import train_test_split
dt_train,dt_test = train_test_split(df,test_size=0.2)
dt_train = dt_train.reset_index()
dt_test = dt_test.reset_index()

create_tf_data_file(dt_train,train_output_file)
create_tf_data_file(dt_test,test_output_file)

9%2

text1 ='夜場所口'
text2 ='夜場所口'
tokenizer.encode_plus(f"{text1}[SEP]{text2}",padding='max_length',max_length=max_seq_length*2,truncation=True)

len([2, 594, 335, 14, 98, 843, 291, 7, 287, 53, 581, 1582, 46, 28355, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0])

len([1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0])

dt_train.shape

dt_test.shape
