==> 04-create-model-texts-similarity.ipynb <==
import pickle
from transformers import BertJapaneseTokenizer, BertForMaskedLM
import pandas as pd
import json
import matplotlib.pyplot as plt
import tensorflow as tf
import seaborn as sns
import collections
from transformers import DistilBertTokenizer
from transformers import TFDistilBertForSequenceClassification
from transformers import DistilBertConfig

max_seq_length = 28

model = TFDistilBertForSequenceClassification.from_pretrained('cl-tohoku/bert-base-japanese-whole-word-masking', num_labels=2)

class MyModel(tf.keras.Model):
    def __init__(self,model):
        super().__init__()
        #self.model = model.distilbert(input_ids,input_mask)[0]
        self.model = model
        self.model.trainable = False
        self.layersG = tf.keras.layers.GlobalMaxPool1D()
        self.layers1 = tf.keras.layers.Dense(100, activation="relu")
        self.layersD2 = tf.keras.layers.Dropout(0.2)
        self.layers3 = tf.keras.layers.Dense(200, activation="relu")
        self.layersD4 = tf.keras.layers.Dropout(0.2)
        self.layers5 = tf.keras.layers.Dense(100, activation="relu")
        self.layersD6 = tf.keras.layers.Dropout(0.2)
        self.layers7 = tf.keras.layers.Dense(50, activation="relu")
        self.layersD8 = tf.keras.layers.Dropout(0.2)
        self.layers9 = tf.keras.layers.Dense(2, activation="softmax")
        
        self.Bidirectional=tf.keras.layers.Bidirectional(tf.keras.layers.LSTM(100, return_sequences=True, dropout=0.1, recurrent_dropout=0.1))
    def call(self, inputs):
        input_ids1 = inputs["input_ids1"]
        input_mask1 = inputs["input_mask1"]
        embedding_layer1  = self.model.distilbert(input_ids1,input_mask1)[0]
        input_ids2 = inputs["input_ids2"]
        input_mask2 = inputs["input_mask2"]        
        embedding_layer2  = self.model.distilbert(input_ids2,input_mask2)[0]
        embedding_layer = tf.concat([embedding_layer1,embedding_layer2],axis = 1)
        X =self.Bidirectional(embedding_layer)
        X = self.layersG(X)
        X = self.layers1(X)
        X = self.layersD2(X)
        X = self.layers3(X)
        X = self.layersD4(X)
        X = self.layers5(X)
        X = self.layersD6(X)
        X = self.layers7(X)
        X = self.layersD8(X)
        X = self.layers9(X)        
        return X

train_output_file = "../input/text-data-similarity/texts_train_data.tfrecord"
test_output_file = "../input/text-data-similarity/texts_test_data.tfrecord"

def select_data_and_label_from_record(record):
    x = {
        "input_ids1": record["input_ids1"],
        "input_mask1": record["input_mask1"],
        "input_ids2": record["input_ids2"],
        "input_mask2": record["input_mask2"],
        #"segment_ids": record["segment_ids"],
    }
    y = record["label_ids"]
    return (x, y)
def _decode_record(record, name_to_features):
    """Decodes a record to a TensorFlow example."""
    return tf.io.parse_single_example(record, name_to_features)

def create_train_test_data(file_name,isTrain = False):
    dataset = tf.data.TFRecordDataset(file_name)
    if isTrain :
        dataset = dataset.repeat(500)
    dataset = dataset.prefetch(tf.data.experimental.AUTOTUNE)
    name_to_features = {
            "input_ids1": tf.io.FixedLenFeature([max_seq_length], tf.int64),
            "input_mask1": tf.io.FixedLenFeature([max_seq_length], tf.int64),
            "input_ids2": tf.io.FixedLenFeature([max_seq_length], tf.int64),
            "input_mask2": tf.io.FixedLenFeature([max_seq_length], tf.int64),
            #"segment_ids": tf.io.FixedLenFeature([max_seq_length], tf.int64),
            "label_ids": tf.io.FixedLenFeature([], tf.int64),
        }
    drop_remainder=False
    dataset = dataset.apply(
        tf.data.experimental.map_and_batch(
            lambda record: _decode_record(record, name_to_features),
            batch_size=1000,
            drop_remainder=drop_remainder,
            num_parallel_calls=tf.data.experimental.AUTOTUNE,
        )
    )
    dataset.cache()
    re_dataset = dataset.map(select_data_and_label_from_record)
    return re_dataset

train_dataset = create_train_test_data(train_output_file,True)
test_dataset = create_train_test_data(test_output_file)

test_dataset

mode2 = MyModel(model)

learning_rate = 1e-3
epsilon = 1e-6
loss = tf.keras.losses.SparseCategoricalCrossentropy()
metric = tf.keras.metrics.SparseCategoricalAccuracy("accuracy")
optimizer = tf.keras.optimizers.Adam(learning_rate=learning_rate, epsilon=epsilon)
mode2.compile(optimizer=optimizer, loss=loss, metrics=[metric])
history = mode2.fit(
    train_dataset,
    steps_per_epoch = 20,
    validation_data=test_dataset,
    validation_steps=4,
    #batch_size=50,
    #shuffle=True,
    epochs=200)
